Change Dev::UI::InteractivePrompt to Dev::UI::Prompt::InteractiveOptions
